gps: Collapse LockWithImports into Lock
There was no real need to delineate between Lock and LockWithImports.
The old Lock had the InputsDigest concept, which was even less feasible
for theoretical implementations of Lock to have, so this can't possibly
be more harmful.
