+++ gps/source_cache_bolt_test.go
@@ -54,11 +54,11 @@ func TestBoltCacheTimeout(t *testing.T) {
 
 	lock := &safeLock{
 		p: []LockedProject{
-			NewLockedProject(mkPI("github.com/sdboyer/gps"), NewVersion("v0.10.0"), []string{"gps"}),
-			NewLockedProject(mkPI("github.com/sdboyer/gps2"), NewVersion("v0.10.0"), nil),
-			NewLockedProject(mkPI("github.com/sdboyer/gps3"), NewVersion("v0.10.0"), []string{"gps", "flugle"}),
-			NewLockedProject(mkPI("foo"), NewVersion("nada"), []string{"foo"}),
-			NewLockedProject(mkPI("github.com/sdboyer/gps4"), NewVersion("v0.10.0"), []string{"flugle", "gps"}),
+			NewLockedProject(mkPI("github.com/sdboyer/gps"), NewVersion("v0.10.0").Pair("foo"), []string{"gps"}),
+			NewLockedProject(mkPI("github.com/sdboyer/gps2"), NewVersion("v0.10.0").Pair("bar"), nil),
+			NewLockedProject(mkPI("github.com/sdboyer/gps3"), NewVersion("v0.10.0").Pair("baz"), []string{"gps", "flugle"}),
+			NewLockedProject(mkPI("foo"), NewVersion("nada").Pair("zero"), []string{"foo"}),
+			NewLockedProject(mkPI("github.com/sdboyer/gps4"), NewVersion("v0.10.0").Pair("qux"), []string{"flugle", "gps"}),
 		},
 	}
 
@@ -236,8 +236,9 @@ func TestBoltCacheTimeout(t *testing.T) {
 
 	newLock := &safeLock{
 		p: []LockedProject{
-			NewLockedProject(mkPI("github.com/sdboyer/gps"), NewVersion("v1"), []string{"gps"}),
+			NewLockedProject(mkPI("github.com/sdboyer/gps"), NewVersion("v1").Pair("rev1"), []string{"gps"}),
 		},
+		i: []string{"foo", "bar"},
 	}
 
 	newPtree := pkgtree.PackageTree{
@@ -285,8 +286,8 @@ func TestBoltCacheTimeout(t *testing.T) {
 		}
 		compareManifests(t, newManifest, gotM)
 		// TODO(sdboyer) use DiffLocks after refactoring to avoid import cycles
-		if !locksAreEq(lock, gotL) {
-			t.Errorf("locks are different:\n\t(GOT): %s\n\t(WNT): %s", lock, gotL)
+		if !locksAreEq(newLock, gotL) {
+			t.Errorf("locks are different:\n\t(GOT): %s\n\t(WNT): %s", newLock, gotL)
 		}
 
 		got, ok := c.getPackageTree(rev, root)
